+++ package.json
@@ -10,7 +10,6 @@
     "@types/node": "^16.7.13",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
-    "@types/sass": "^1.43.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
@@ -60,7 +59,8 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "husky": ">=6",
     "lint-staged": ">=10",
-    "prettier": "^2.7.1"
+    "prettier": "^2.7.1",
+    "@types/sass": "^1.43.1"
   },
   "lint-staged": {
     "*.{ts,tsx}": [
