+++ src/types/movie.ts
@@ -1,0 +1,10 @@
+export interface Movie {
+  id: number;
+  img: string;
+  title: string;
+  year: number;
+  rating: number;
+  duration: string;
+  shortDescription: string;
+  description: string;
+}
